test: using random string for retry of the tests
