+++ qa/c8-orchestration-cluster-e2e-test-suite/tests/common-flows/identity-users-flows.spec.ts
@@ -11,6 +11,7 @@ import {navigateToApp} from '@pages/UtilitiesPage';
 import {expect} from '@playwright/test';
 import {captureScreenshot, captureFailureVideo} from '@setup';
 import {relativizePath, Paths} from 'utils/relativizePath';
+import {generateRandomStringAsync} from '../../utils/randomString';
 
 test.describe('Users Page Tests', () => {
   test.beforeEach(async ({loginPage, page}) => {
@@ -23,18 +24,22 @@ test.describe('Users Page Tests', () => {
     await captureFailureVideo(page, testInfo);
   });
 
-  const TEST_USER = {
-    username: 'yuliia',
-    password: 'yuliia',
-    email: 'yuliia@example.com',
-  };
-
   test('Admin user can delete user', async ({
     page,
     loginPage,
     identityUsersPage,
     identityHeader,
   }) => {
+    const randomString = await generateRandomStringAsync(3);
+
+    const TEST_USER = {
+      username: 'Test' + randomString,
+      name: 'Test User' + randomString,
+      email: `test${randomString}@test.com`,
+      password: `test${randomString}`,
+    };
+    await identityUsersPage.createUser(TEST_USER);
+
     await expect(identityUsersPage.usersList).toBeVisible();
     await expect(identityUsersPage.usersList).toContainText(TEST_USER.username);
     await identityUsersPage.deleteUser(TEST_USER);
